requirements.txt
Django==5.2.6
djangorestframework==3.14.0
requests==2.31.0
urllib3==2.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9